Refactor batching into its own module
