+++ mlserver/batching/__init__.py
@@ -1,0 +1,3 @@
+from .adaptive import AdaptiveBatcher
+
+__all__ = ["AdaptiveBatcher"]
